CUA/Backend/requirements.txt
openai==1.54.3
python-dotenv==1.0.0
playwright==1.47.0
Pillow==10.1.0
azure-identity==1.15.0
